+++ docs/release-notes.rst
@@ -99,11 +99,13 @@ Deprecated Features
 Bug Fixes
 .........
 
-- Fixed an `issue <https://github.com/square/bionic/issues/111>`_ where non-persistable
+- Fixed an `issue <https://github.com/square/bionic/issues/111>`__ where non-persistable
   entities could be spuriously recomputed even when their values weren't directly
   needed.
-- Fixed an `issue <https://github.com/square/bionic/issues/229>`_ where caching fails
+- Fixed an `issue <https://github.com/square/bionic/issues/229>`__ where caching fails
   when cache directory paths contain whitespaces.
+- Fixed an `issue <https://github.com/square/bionic/issues/233>`__ where reloading of
+  flows that merge other flows from other modules does not work correctly.
 
 Documentation
 .............